changed RandR context to be a shortcut like global
