+++ src/components/App.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { GlobalContextProvider } from '../context/GlobalStore';
-import RatingsAndReviews from './ratings-reviews/RatingsAndReviews';
-import RelatedItems from './related-items/RelatedItems';
+import { RAndRContextProvider } from '../context/RAndRContext';
 import ProductDetail from './product-detail/productDetail';
+import RelatedItems from './related-items/RelatedItems';
+import RatingsAndReviews from './ratings-reviews/RatingsAndReviews';
 
 function App() {
   console.log('app [rerendered]');
@@ -11,7 +12,9 @@ function App() {
       <ProductDetail />
       {/* <QuestionsAndAnswers /> */}
       <RelatedItems />
-      <RatingsAndReviews />
+      <RAndRContextProvider>
+        <RatingsAndReviews />
+      </RAndRContextProvider>
     </GlobalContextProvider>
   );
 }
