+++ src/components/ratings-reviews/RatingsAndReviews.jsx
@@ -1,33 +1,27 @@
-import React, { createContext, useState, useEffect } from 'react';
+import React, { createContext, useState, useEffect, useContext } from 'react';
 import axios from 'axios';
 import Reviews from './reviews/Reviews';
 import Ratings from './ratings/Ratings';
 import { useGlobalContext } from '../../context/GlobalStore';
-
-export const RatingsAndReviewsContext = createContext();
+import { useRAndRContext } from '../../context/RAndRContext';
 
 export default function RatingsAndReviews() {
-  const { productId, setProductId, setAvgRating } = useGlobalContext();
-
-  const [reviewsSort, setReviewsSort] = useState(() => 'relevant');
-  const [showAdd, setShowAdd] = useState(() => true);
-  const [reviewsMeta, setReviewsMeta] = useState(() => { });
-  const [totalRatings, setTotalRatings] = useState(() => 0);
-  const [reviews, setReviews] = useState(() => []);
-  const [page, setPage] = useState(() => 1);
-  const [reviewsFilter, setReviewsFilter] = useState({
-    1: false,
-    2: false,
-    3: false,
-    4: false,
-    5: false
-  });
+  const { productId, setAvgRating } = useGlobalContext();
+  const {
+    setReviewsMeta,
+    setReviews,
+    setShowAdd,
+    setPage,
+    reviewsSort,
+    reviews,
+    page,
+  } = useRAndRContext();
   console.log('ratings and reviews [rendered]');
 
   useEffect(() => {
     axios.get(`/reviews/meta/${productId}`)
       .then(({ data }) => {
-        console.log('meta', data)
+        console.log('meta', data);
         setReviewsMeta(data);
         setAvgRating(data.averageRating);
       })
@@ -64,29 +58,10 @@ export default function RatingsAndReviews() {
       });
   }, [page]);
 
-  const value = {
-    reviewsSort,
-    setReviewsSort,
-    showAdd,
-    setShowAdd,
-    reviewsMeta,
-    setReviewsMeta,
-    totalRatings,
-    setTotalRatings,
-    reviews,
-    setReviews,
-    page,
-    setPage,
-    reviewsFilter,
-    setReviewsFilter,
-  };
-
   return (
-    <RatingsAndReviewsContext.Provider value={value}>
-      <div id="ratings-and-reviews">
-        <Ratings />
-        <Reviews />
-      </div>
-    </RatingsAndReviewsContext.Provider>
+    <div id="ratings-and-reviews">
+      <Ratings />
+      <Reviews />
+    </div>
   );
 }
